Rename buy membership modals
