+++ packages/ui/src/components/page/Sidebar/CurrentMember/CurrentMember.tsx
@@ -5,7 +5,7 @@ import { BorderRad, Colors, Transitions } from '../../../../constants'
 import { useModal } from '../../../../hooks/useModal'
 import { useMyMemberships } from '../../../../hooks/useMyMemberships'
 import { useToggle } from '../../../../hooks/useToggle'
-import { BuyMembershipModalCall } from '../../../../modals/AddMembershipModal'
+import { BuyMembershipModalCall } from '../../../../modals/BuyMembershipModal'
 import { ArrowDownExpandedIcon, Icon } from '../../../icons'
 import { MemberDarkHover, MemberInfo, MembershipsCount } from '../../../membership'
 import { AddMembershipButton } from '../../../membership/AddMembershipButton'
