--- packages/ui/src/modals/AddMembershipModal/index.tsx
@@ -1,2 +1,0 @@
-export * from './AddMembershipModal'
-export * from './types'
