+++ packages/ui/src/modals/BuyMembershipModal/index.tsx
@@ -1,0 +1,2 @@
+export * from './BuyMembershipModal'
+export * from './types'
