+++ packages/ui/src/modals/InviteMemberModal/InviteFormModal.tsx
@@ -20,7 +20,7 @@ import { useFormValidation } from '../../hooks/useFormValidation'
 import { useKeyring } from '../../hooks/useKeyring'
 import { useObservable } from '../../hooks/useObservable'
 import { AvatarURISchema, HandleSchema, MemberSchema, NewAddressSchema } from '../../membership/data/validation'
-import { FormFields, formReducer } from '../AddMembershipModal/formReducer'
+import { FormFields, formReducer } from '../BuyMembershipModal/formReducer'
 import { Row } from '../common'
 
 interface InviteProps {
